# src/routes/swagger.json
{
	"components": {
...
		"schemas": {
			"AcceptableDbms": {
				"enum": [
					"mysql",
					"oracle",
					"mariadb",
					"postgres",
					"cubrid"
				],
				"type": "string"
			},
			"UserInterface": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"token",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"$ref": "#/components/schemas/UserInterface"
			},
			"Service": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"method": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"entityName": {
						"type": "string"
					},
					"tableName": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"application": {
						"$ref": "#/components/schemas/Application"
					},
					"stage": {
						"$ref": "#/components/schemas/Stage"
					},
					"stageId": {
						"type": "number",
						"format": "double"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"columns": {
						"items": {
							"$ref": "#/components/schemas/ServiceColumn"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"method",
					"description",
					"entityName",
					"tableName",
					"status",
					"user",
					"application",
					"stage",
					"stageId",
					"meta",
					"columns",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Application": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nameSpace": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"trafficConfigs": {
						"items": {
							"$ref": "#/components/schemas/TrafficConfig"
						},
						"type": "array"
					},
					"stages": {
						"items": {
							"$ref": "#/components/schemas/Stage"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"nameSpace",
					"title",
					"description",
					"userId",
					"user",
					"services",
					"trafficConfigs",
					"stages",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrafficConfig": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"application": {
						"$ref": "#/components/schemas/Application"
					},
					"type": {
						"type": "string"
					},
					"maxCount": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"application",
					"type",
					"maxCount",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Stage": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"application": {
						"$ref": "#/components/schemas/Application"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status",
					"application",
					"services",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Meta": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"dataType": {
						"type": "string"
					},
					"originalFileName": {
						"type": "string"
					},
					"remoteFilePath": {
						"type": "string"
					},
					"filePath": {
						"type": "string"
					},
					"encoding": {
						"type": "string"
					},
					"extension": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"port": {
						"type": "string"
					},
					"db": {
						"type": "string"
					},
					"dbUser": {
						"type": "string"
					},
					"pwd": {
						"type": "string"
					},
					"table": {
						"type": "string"
					},
					"dbms": {
						"$ref": "#/components/schemas/AcceptableDbms"
					},
					"rowCounts": {
						"type": "number",
						"format": "double"
					},
					"skip": {
						"type": "number",
						"format": "double"
					},
					"sheet": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					},
					"columns": {
						"items": {
							"$ref": "#/components/schemas/MetaColumn"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"dataType",
					"originalFileName",
					"remoteFilePath",
					"filePath",
					"encoding",
					"extension",
					"host",
					"port",
					"db",
					"dbUser",
					"pwd",
					"table",
					"dbms",
					"rowCounts",
					"skip",
					"sheet",
					"isActive",
					"user",
					"service",
					"columns",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceColumn": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"columnName": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"hidden": {
						"type": "boolean"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"columnName",
					"type",
					"hidden",
					"service",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AcceptableType": {
				"enum": [
					"bigint",
					"int",
					"double",
					"bit",
					"date",
					"datetime",
					"time",
					"varchar",
					"text",
					"longtext",
					"boolean"
				],
				"type": "string"
			},
			"MetaColumn": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"originalColumnName": {
						"type": "string"
					},
					"columnName": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/AcceptableType"
					},
					"originalType": {
						"$ref": "#/components/schemas/AcceptableType"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"isHidden": {
						"type": "boolean"
					},
					"isSearchable": {
						"type": "boolean"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"params": {
						"items": {
							"$ref": "#/components/schemas/MetaParam"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"originalColumnName",
					"columnName",
					"type",
					"originalType",
					"size",
					"order",
					"isHidden",
					"isSearchable",
					"meta",
					"params",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ParamOperatorType": {
				"enum": [
					"lt",
					"lte",
					"gt",
					"gte",
					"like",
					"eq",
					"neq"
				],
				"type": "string"
			},
			"MetaParam": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"metaColumn": {
						"$ref": "#/components/schemas/MetaColumn"
					},
					"operator": {
						"$ref": "#/components/schemas/ParamOperatorType"
					},
					"description": {
						"type": "string"
					},
					"isRequired": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"metaColumn",
					"operator",
					"description",
					"isRequired",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DbmsParams": {
				"properties": {
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"dbms": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"port": {
						"type": "string"
					},
					"database": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"table": {
						"type": "string"
					}
				},
				"required": [
					"serviceId",
					"title",
					"dbms",
					"host",
					"port",
					"database",
					"user",
					"password",
					"table"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FileParams": {
				"properties": {
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"dataType": {
						"type": "string"
					},
					"ext": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"skip": {
						"type": "number",
						"format": "double"
					},
					"sheet": {
						"type": "number",
						"format": "double"
					},
					"filePath": {
						"type": "string"
					},
					"originalFileName": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"serviceId",
					"dataType",
					"ext",
					"title",
					"skip",
					"sheet"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DatabaseConnection": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"connectionName": {
						"type": "string"
					},
					"hostname": {
						"type": "string"
					},
					"port": {
						"type": "string"
					},
					"database": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"dbms": {
						"$ref": "#/components/schemas/AcceptableDbms"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"connectionName",
					"hostname",
					"port",
					"database",
					"username",
					"password",
					"dbms",
					"user",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DatabaseConnectionCreateParams": {
				"properties": {
					"title": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"port": {
						"type": "string"
					},
					"db": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"pwd": {
						"type": "string"
					},
					"dbms": {
						"$ref": "#/components/schemas/AcceptableDbms"
					}
				},
				"required": [
					"title",
					"host",
					"port",
					"db",
					"user",
					"pwd",
					"dbms"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceParams": {
				"properties": {
					"method": {
						"type": "string"
					},
					"entityName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"applicationId": {
						"type": "number",
						"format": "double"
					},
					"fileParams": {
						"$ref": "#/components/schemas/FileParams"
					},
					"dbmsParams": {
						"$ref": "#/components/schemas/DbmsParams"
					}
				},
				"required": [
					"method",
					"entityName",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"scopes": {}
			}
		}
	},
	"info": {
		"title": "api-gen",
		"version": "0.0.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/database-connections": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"perPage": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"totalCount": {
											"type": "number",
											"format": "double"
										},
										"dbcs": {
											"items": {
												"$ref": "#/components/schemas/DatabaseConnection"
											},
											"type": "array"
										}
									},
									"required": [
										"perPage",
										"page",
										"totalCount",
										"dbcs"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Database Connection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Post",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseConnection"
								}
							}
						}
					}
				},
				"tags": [
					"Database Connection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DatabaseConnectionCreateParams"
							}
						}
					}
				}
			}
		},
		"/api/database-connections/{connectionId}": {
			"get": {
				"operationId": "GetTablesInConnection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "DB에 접속하여 Table 목록을 불러옵니다.",
				"tags": [
					"Database Connection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Database Connection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/database-connections/{connectionId}/tables": {
			"get": {
				"operationId": "GetConnection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Database Connection의 상세 정보를 보여줍니다.",
				"tags": [
					"Database Connection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Database Connection의 id",
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/database-connections/{connectionId}/tables/{tableName}": {
			"get": {
				"operationId": "GetTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "DB에 접속하여 Table의 컬럼 목록과 정보를 가져옵니다.",
				"tags": [
					"Database Connection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "정보를 가져오려는 테이블 명",
						"in": "path",
						"name": "tableName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/services": {
			"post": {
				"operationId": "Post",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Service"
								}
							}
						}
					}
				},
				"tags": [
					"Service"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ServiceParams"
							}
						}
					}
				}
			}
		},
		"/api/services/{serviceId}": {
			"put": {
				"operationId": "Put",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Service"
								}
							}
						}
					}
				},
				"tags": [
					"Service"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "serviceId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ServiceParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Service"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "serviceId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/stages/{id}/deploy": {
			"post": {
				"operationId": "Deploy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Stage"
								}
							}
						}
					}
				},
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/stages/{id}/undeploy": {
			"post": {
				"operationId": "Undeploy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Stage"
								}
							}
						}
					}
				},
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/stages/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/rest/services": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"#Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "entityName",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rest/services/{id}": {
			"get": {
				"operationId": "GetDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"#Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}